requirements.txt
pyzmq
pyserial
pyroute2
wheel
psutil
connexion
flask-cors
swagger_ui_bundle
